MY PORTFOLIO
Welcome to my portfolio! I'm an expert software engineer proficient in a range of programming languages, including HTML, CSS, JavaScript, React, MQL4 and 5, and NinjaTrader. With a knack for both front-end and back-end development, I specialize in creating dynamic and engaging web applications. Whether you need a sleek and responsive user interface or a robust and efficient backend system, I have the skills and experience to bring your ideas to life. Explore my portfolio to see some of my projects and get in touch to discuss how we can collaborate on your next venture!